+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Create/delete and list Lambda functions
 - Create/delete and list autoscaling launch configurations and autoscaling groups
 - List EC2 images (AMIs) of which you are the owner: `awless list images`
+- Copy an EC2 image from a given region to the current region: `awless copy image name=... source-id=... source-region=...`
 - List your IAM access keys: `awless list accesskeys`
 - Friendlier (using units: KB, MB, GB) display for storage size (s3objects, volumes, lambda functions)